// src/types/bookmarks.d.ts
export interface BookmarkData {
  title: string;
  creator: string;
  url: string;
  tags: string[];
}
// ...
export interface BookmarkingResponse {
  success: boolean;
  message: string;
  source: string;
}